+++ tests/processing/test_processor.py
@@ -8,6 +8,7 @@ from arroyo.backends.local.backend import LocalBroker
 from arroyo.backends.local.storages.abstract import MessageStorage
 from arroyo.backends.local.storages.memory import MemoryMessageStorage
 from arroyo.commit import IMMEDIATE, CommitPolicy
+from arroyo.dlq import DlqPolicy, InvalidMessage
 from arroyo.processing.processor import InvalidStateError, StreamProcessor
 from arroyo.processing.strategies.abstract import (
     MessageRejected,
@@ -122,13 +123,26 @@ def test_stream_processor_lifecycle() -> None:
     with assert_changes(lambda: int(consumer.close.call_count), 0, 1):
         processor._shutdown()
 
-    poll_metric, processing_metric, pause_metric = metrics.calls
+    (
+        poll_metric,
+        callback_metric,
+        processing_metric,
+        pause_metric,
+        join_metric,
+        shutdown_metric,
+    ) = metrics.calls
     assert isinstance(poll_metric, Timing)
     assert poll_metric.name == "arroyo.consumer.poll.time"
+    assert isinstance(callback_metric, Timing)
+    assert callback_metric.name == "arroyo.consumer.callback.time"
     assert isinstance(processing_metric, Timing)
     assert processing_metric.name == "arroyo.consumer.processing.time"
     assert isinstance(pause_metric, Timing)
     assert pause_metric.name == "arroyo.consumer.paused.time"
+    assert isinstance(join_metric, Timing)
+    assert join_metric.name == "arroyo.consumer.join.time"
+    assert isinstance(shutdown_metric, Timing)
+    assert shutdown_metric.name == "arroyo.consumer.shutdown.time"
 
 
 def test_stream_processor_termination_on_error() -> None:
@@ -162,6 +176,91 @@ def test_stream_processor_termination_on_error() -> None:
         processor.run()
 
     assert e.value == exception
+
+
+def test_stream_processor_invalid_message_from_poll() -> None:
+    topic = Topic("test")
+
+    consumer = mock.Mock()
+    partition = Partition(topic, 0)
+    offset = 1
+    now = datetime.now()
+
+    consumer.poll.side_effect = [BrokerValue(0, partition, offset, now)]
+
+    strategy = mock.Mock()
+    strategy.poll.side_effect = [InvalidMessage(partition, 0, needs_commit=False), None]
+
+    factory = mock.Mock()
+    factory.create_with_partitions.return_value = strategy
+
+    processor: StreamProcessor[int] = StreamProcessor(
+        consumer, topic, factory, IMMEDIATE
+    )
+
+    assignment_callback = consumer.subscribe.call_args.kwargs["on_assign"]
+    assignment_callback({Partition(topic, 0): 0})
+
+    # We need to ensure that poll() is called again after it raises
+    # InvalidMessage the first time. This gives the strategy time to commit
+    # offsets for that message before the next one is submitted.
+    processor._run_once()
+    assert strategy.poll.call_count == 1
+
+    processor._run_once()
+    assert strategy.poll.call_count == 2
+    assert strategy.submit.call_count == 1
+
+
+def test_stream_processor_invalid_message_from_submit() -> None:
+    topic = Topic("test")
+
+    consumer = mock.Mock()
+    partition = Partition(topic, 0)
+    offset = 1
+    now = datetime.now()
+
+    consumer.poll.side_effect = [
+        BrokerValue(0, partition, offset, now),
+        BrokerValue(1, partition, offset + 1, now),
+    ]
+
+    strategy = mock.Mock()
+    strategy.submit.side_effect = [
+        InvalidMessage(partition, 0, needs_commit=False),
+        None,
+        None,
+    ]
+
+    factory = mock.Mock()
+    factory.create_with_partitions.return_value = strategy
+
+    processor: StreamProcessor[int] = StreamProcessor(
+        consumer, topic, factory, IMMEDIATE
+    )
+
+    assignment_callback = consumer.subscribe.call_args.kwargs["on_assign"]
+    assignment_callback({Partition(topic, 0): 0})
+
+    processor._run_once()
+    assert strategy.poll.call_count == 1
+    assert consumer.pause.call_count == 0
+    assert strategy.submit.call_count == 1
+
+    processor._run_once()
+    assert strategy.submit.call_count == 2
+    assert strategy.submit.call_args_list == [
+        mock.call(Message(BrokerValue(0, partition, offset, now))),
+        mock.call(Message(BrokerValue(0, partition, offset, now))),
+    ]
+
+    processor._run_once()
+    assert strategy.submit.call_count == 3
+    assert strategy.submit.call_args_list == [
+        mock.call(Message(BrokerValue(0, partition, offset, now))),
+        mock.call(Message(BrokerValue(0, partition, offset, now))),
+        mock.call(Message(BrokerValue(1, partition, offset + 1, now))),
+    ]
 
 
 def test_stream_processor_create_with_partitions() -> None:
@@ -429,3 +528,31 @@ def test_commit_policy_bench(
             processor._run_once()
 
     benchmark(inner)
+
+
+def test_dlq() -> None:
+    topic = Topic("topic")
+    partition = Partition(topic, 0)
+    consumer = mock.Mock()
+    consumer.poll.return_value = BrokerValue(0, partition, 1, datetime.now())
+    strategy = mock.Mock()
+    strategy.submit.side_effect = InvalidMessage(partition, 1)
+    factory = mock.Mock()
+    factory.create_with_partitions.return_value = strategy
+
+    dlq_policy: Any = DlqPolicy(producer=mock.Mock())
+
+    processor: StreamProcessor[int] = StreamProcessor(
+        consumer, topic, factory, IMMEDIATE, dlq_policy
+    )
+
+    # Assignment
+    subscribe_args, subscribe_kwargs = consumer.subscribe.call_args
+    assert subscribe_args[0] == [topic]
+    assignment_callback = subscribe_kwargs["on_assign"]
+    offsets = {Partition(topic, 0): 0}
+    assignment_callback(offsets)
+
+    processor._run_once()
+
+    assert dlq_policy.producer.produce.call_count == 1